packages/auto-scale/main/requirements.txt
kubernetes==31.0.0
pyyaml==6.0.2
aiohttp==3.11.10